Updates to fix error with closed sockets.
The issue is that if the GMetric.announce method is used multiple times
then the underlying socket in the protocol will be closed on the second
(and subsequent) calls. The changes in this commit use the Closeable
interface to mark the GMetric and Protocol classes as resources that
should be closed. Eclipse and other IDEs should warn client code if
these resources are not being properly closed. This design allows usage
of the GMetric class as either one time usage or with multiple calls.

- Added Closeable interface to Protocol and GMetric along with
implementation.
- Updated constructors of GMetric and Protocol implementations to throw
IOException.
- Updated AbstractProtocol to create the socket in the
constructor.
- Updated POM version to 1.0.3-SNAPSHOT.
- Removed unused imports from GMonitor and GSampler.
- Removed usage of @Override for interface implementation methods which
is an error in 1.5 compliance mode.